Data.Char.Properties.XMLCharProps

Description

XML character properties

Documentation

isXmlChar :: Char -> Bool

checking for valid XML characters

isXmlCharCR :: Char -> Bool

checking for valid XML characters, except CR

isXml1ByteChar :: Char -> Bool

check for a legal 1 byte XML char

isXmlLatin1Char :: Char -> Bool

test for a legal latin1 XML char

isXmlSpaceChar :: Char -> Bool

checking for XML space character: \n, \r, \t and " "

isXmlSpaceCharCR :: Char -> Bool

checking for XML space character: \n, \t and " "

isXml11SpaceChar :: Char -> Bool

checking for XML1.1 space character: additional space 0x85 and 0x2028

see also : isXmlSpaceChar

isXmlNameChar :: Char -> Bool

checking for XML name character

isXmlNameStartChar :: Char -> Bool

checking for XML name start character

see also : isXmlNameChar

isXmlNCNameChar :: Char -> Bool

checking for XML NCName character: no ":" allowed

see also : isXmlNameChar

isXmlNCNameStartChar :: Char -> Bool

checking for XML NCName start character: no ":" allowed

see also : isXmlNameChar, isXmlNCNameChar

isXmlPubidChar :: Char -> Bool

checking for XML public id character

isXmlLetter :: Char -> Bool

checking for XML letter

isXmlBaseChar :: Char -> Bool

checking for XML base charater

isXmlIdeographicChar :: Char -> Bool

checking for XML ideographic charater

isXmlCombiningChar :: Char -> Bool

checking for XML combining charater

isXmlDigit :: Char -> Bool

checking for XML digit

isXmlExtender :: Char -> Bool

checking for XML extender

isXmlControlOrPermanentlyUndefined :: Char -> Bool

checking for XML control or permanently discouraged char

see Errata to XML1.0 (http://www.w3.org/XML/xml-V10-2e-errata) No 46

Document authors are encouraged to avoid "compatibility characters", as defined in section 6.8 of [Unicode] (see also D21 in section 3.6 of [Unicode3]). The characters defined in the following ranges are also discouraged. They are either control characters or permanently undefined Unicode characters:

charPropXmlChar :: CharSet

charPropXmlCharCR :: CharSet

charPropXml1ByteChar :: CharSet

charPropXmlLatin1Char :: CharSet

charPropXmlSpaceChar :: CharSet

charPropXmlSpaceCharCR :: CharSet

charPropXml11SpaceChar :: CharSet

charPropXmlNameChar :: CharSet

charPropXmlNameStartChar :: CharSet

charPropXmlNCNameChar :: CharSet

charPropXmlNCNameStartChar :: CharSet

charPropXmlPubidChar :: CharSet

charPropXmlLetter :: CharSet

charPropXmlBaseChar :: CharSet

charPropXmlIdeographicChar :: CharSet

charPropXmlCombiningChar :: CharSet

charPropXmlDigit :: CharSet

charPropXmlExtender :: CharSet

charPropXmlControlOrPermanentlyUndefined :: CharSet